# docs/openapi/tessera.yaml
...
openapi: 3.1.0
info:
  title: Tessera viewer and session planes
  version: "1"
  summary: A permission-masked point service — the two planes a client speaks to.
  description: |
    Two listeners. The **session plane** mints and revokes viewer sessions and is authenticated by
    the operator-configured *session credential*, a shared secret; it is never browser-facing. The
    **viewer plane** answers about the corpus as one principal may see it and is authenticated by
    the *session token* the session plane minted. `/healthz` and `/readyz` sit on both, unauthenticated,
    bare status codes with no body.

    `api_version` is published by `GET /v1/meta` and by nothing else. There is no version header,
    on either side, so a client must not send one as a negotiation and must not expect a refusal
    if its assumed version is wrong (contracts §3.1).

    **Every error is one envelope** (`Error` below) with a closed code list, and the codes group
    into three refusal classes that tell a client whether to retry (contracts §3.1, design §10.6):

    - **Failure** — `500 fail-closed`, `503 not-ready`. Fail-closed and never a partial result.
    - **Backpressure** — `429 backpressure`. Retry the request *unchanged*, after `Retry-After`,
      which every 429 carries together with an agreeing body `retry_after_s`. The number is
      per-subject — `1` for the viewer and session planes' compute-admission gate, derived
...

...

...
servers:
  - url: http://{viewer-host}
...
    variables:
...
paths:
...
  /v1/viewport:
    post:
      tags: [viewer]
      operationId: viewport
      summary: Masked counts per tile, a served sample of points, and the served annotation artifacts, over a region at a depth.
      description: |
        The response is **not JSON**. It is a sequence of tagged, length-prefixed frames —
        `u8 kind`, `u32 little-endian payload length`, payload — each payload a complete Arrow IPC
        stream (JSON for the trailer), decodable alone (contracts §5). Kinds: `1` tiles (exactly
        one, first), `2` sub-cells (exactly one iff `underlay_offset` was supplied and non-zero),
        `5` artifacts (at most one, after tiles and before any points; **absent, never empty, when
        nothing is served**, deliberately), `3` points (zero or more, whole tiles per frame,
        boundaries not contract), `4` trailer (exactly one, last — its presence is the completeness
        signal). A body without a trailer is incomplete whatever the transport said; an unknown
        kind is a decoder error, never skipped. `docs/openapi/README.md` walks a decode in Python
        and JavaScript, and contracts §3.2 gives every batch's schema.

        **Exactly one of `bbox` and `tiles`.** `k` defaults to the deployment's `k_max_marks` and
        is capped by `max_k` (both in `/v1/meta`'s `selection`); it is design §7.2's cap clause,
        so a tile may return fewer than `k` while more are visible — that is the density signal,
        not truncation. **`k = 0` is the counts-only request**: tiles and artifacts as at any `k`,
        and no points frame at all. **A client must not decrease `k` as it zooms in**; the server
        cannot enforce this.

        The *artifacts* frame's `matched` column answers the request's `filters` per
        artifact, and is null where the request carried none — see `filters`. Its `layer` column
        is **dictionary-encoded** (u16 keys over utf8 values); its `rung` column is the number a
        client draws by — the declared level on a levelled layer, the response-local parent-chain
        depth on a treed one (the depth in the forest this response's own `parent_id` links form,
        after the budget cut), `0` on a flat one; and its two shape columns `shape_x`/`shape_y` —
        `list<list<list<uint32>>>`, parts then rings then vertices, one drawn geometry per artifact
        of the kind its layer's `shape` names — **trail the fixed columns and are absent from the
        schema entirely when no served layer draws a shape or the request did not ask for one** —
        an absent column is distinguishable from a null one, so a per-row null keeps its one
        meaning: *this layer declares no such property*, never *withheld* (decision 0076).
        `artifact_rows: "identity"` answers with the **same rows** in a fixed five-column schema —
        see its property.

        A request carrying a `highlight` gains three columns and loses none: `highlighted` fifth on
        the tiles frame (`uint64`, **always present**, equal to `matched` where the request carried
        no highlight), `highlighted` on the points frame (`bool`, after the render scalars and
        before the membership columns, **present only with a highlight**), and `highlighted`
        fifteenth on the artifacts frame (`bool?`, after `matched`). The served set is identical
        with and without one — see `highlight`.

        The artifacts frame's sixteenth column is `target` (`uint64?`): the `tessera_id` of the
        artifact this row is attached to, which for a clustering's label is its cluster. The
        target is a row of the same frame, and the column is null for a row attached to nothing.
        A label's `masked_count` is its own. The `"identity"` projection omits `target`.

        `layers` decides whether an artifacts frame can appear — see its schema; the semantics
        there are the owner's ruling of 2026-08-25 and the server change is landing on the s3
        track, so until that merges an *omitted* `layers` is read as `"all"` and the string form is
        not accepted. `artifact_budget` is honoured structurally (ancestors in place of their
        descendants), never by sampling, and is inert on a flat layer.

        Behind the compute-admission gate.
      security:
        - sessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewportRequest"
      responses:
        "200":
          description: The framed body. Headers carry the coordinates a client caches and revalidates by.
          headers:
            etag:
              description: The **content key**, quoted — identity, the served watermark, the overlay version, a per-process nonce (`delta-serving.md` §2). What a held band may be *declared* under. Echoed as `If-Match`; a mismatch yields the full response, never `412`.
              required: true
              schema:
                type: string
            x-tessera-identity-key:
              description: The **identity key** — idset, auth-data hash, mask-fragment identity, view. What a held band may be *rendered* under at all; the client's cache partition key.
              required: true
              schema:
                type: string
            x-tessera-pin:
              description: The generation stamp of the geometry this response was computed from, as JSON `{"prefix", "segments_version"}`. A stamp, not a selector — echo it back as `pin`, and the only effect is `x-tessera-stale`.
              required: true
              schema:
                type: string
            x-tessera-stale:
              description: "`1` if geometry has moved since the `pin` this request presented, else `0`. The **broadcast** form: it moves at every flush, merge and compaction whether or not anything this principal can see changed. Advisory in both directions."
              required: true
              schema:
                type: string
                enum: ["0", "1"]
            x-tessera-region:
              description: "Present exactly when the request's `filters` carried a `region` leaf: `exact` when every region's answer is exact for the shape against each point's stored position, or `cover; depth=<d>` when a shape's perimeter exceeded `max_region_cells` and the answer is exact for a cover of it — a superset — taken at depth `d` (`selection-operand.md` §6). A function of the shape and the grid alone, never of the rows, so it says nothing about the corpus; with several region leaves, the coarsest. A leaf by artifact is always exact."
              required: false
              schema:
                type: string
                pattern: "^(exact|cover; depth=[0-9]+)$"
            x-tessera-server-us:
              description: Server compute after admission to first-flush-ready, microseconds. The whole-stream wall total is the trailer's `stream_us`.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
            x-tessera-admission-us:
              description: Time spent in the compute-admission gate, microseconds.
              required: true
              schema:
                type: string
                pattern: "^[0-9]+$"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Framed Arrow IPC streams and one JSON trailer, per contracts §5.
        "401":
          $ref: "#/components/responses/BadCredential"
        "403":
          $ref: "#/components/responses/ExpiredToken"
        "404":
          description: Unknown `view`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "`zoom` above 16; neither or both of `bbox`/`tiles`; a non-finite or inverted bbox; a tile prefix with bits above `zoom`; an `underlay_offset` beyond the deployment's ceiling, the grid or the sub-cell budget (refused, never clamped); an unknown filter column or an operator outside its family (an unknown *value* is an empty operand, never a refusal); `none_of` naming a text column or naming two columns."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Backpressure"
        "500":
          $ref: "#/components/responses/FailClosed"
        "503":
          $ref: "#/components/responses/NotReady"
...
  /v1/artifacts/{tessera_id}:
    post:
      tags: [viewer]
      operationId: artifact
      summary: Drill down on one annotation artifact — its masked count and derived geometry for this principal.
      description: |
        A separate route from `/v1/items`, so that the *shape* of an answer never says which of the
        two an identifier names. `masked_count` is **this principal's** count of the artifact's
        members and never its size. Geometry is recomputed per principal from `membership ∩ M_auth`
        and travels in the same 32-bit-per-axis grid units as the viewport's `code`; a client must
        not cache one principal's geometry against an identifier. An absent geometry field means
        *this layer declares no such property*, never *withheld*: an artifact whose content cannot
        be served is a `404` entire.

        `404` is the only failure shape and one construction site: nothing by that id, a point by
        that id, a layer this principal cannot reach, a suppressed artifact, and one below its
        layer's existence criterion are one answer. Behind the compute-admission gate.
      security:
        - sessionToken: []
      parameters:
        - $ref: "#/components/parameters/TesseraId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactRequest"
      responses:
        "200":
          description: The artifact as this principal sees it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactResponse"
        "401":
          $ref: "#/components/responses/BadCredential"
        "403":
          $ref: "#/components/responses/ExpiredToken"
        "404":
          description: "`404 unknown`: nothing this principal may see by that identifier — and an unknown `view` takes the same answer, since a view is addressed like any other named thing."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Stale `idset`, on `/v1/items`' rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Backpressure"
        "500":
          $ref: "#/components/responses/FailClosed"
components:
  securitySchemes:
    sessionCredential:
      type: http
      scheme: bearer
      description: The operator-configured session credential (`serve.session_credential`), a shared secret. Compared in constant time. Never handed to a browser.
    sessionToken:
      type: http
      scheme: bearer
      description: The `token` `POST /session/authorise` minted. A missing, malformed or unrecognised token — including a revoked one — is `401 bad-credential`; a recognised token past `expires_at` is `403 expired-token`. Both end the session.
  parameters:
    TesseraId:
      name: tessera_id
      in: path
      required: true
      description: The item's or artifact's opaque identifier, a `u64` in decimal. A keyed blinding permutation of the internal identity — stable across sessions, meaningless without the key, and **not** a cryptographic guarantee (contracts §5, decision 0014).
      schema:
        type: string
        pattern: "^[0-9]+$"
  responses:
    BadCredential:
      description: "`401 bad-credential`: missing, malformed or unrecognised bearer credential. On a token that previously worked this is the session ending — revoked — and a client treats it exactly as `403`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ExpiredToken:
      description: "`403 expired-token`: the token is recognised and its `expires_at` has passed. Re-authorise."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Backpressure:
      description: "`429 backpressure`: shed whole, never a narrowed answer. Retry unchanged after `Retry-After`, which agrees with the body's `retry_after_s`."
      headers:
        Retry-After:
          required: true
          description: Seconds. `1` for the compute-admission gate and the single-flight builders; other subjects derive their own figure. Read it rather than assuming it.
          schema:
            type: string
            pattern: "^[0-9]+$"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BackpressureError"
    FailClosed:
      description: "`500 fail-closed`: a mask, composition or containment failure. Partial results do not exist."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotReady:
      description: "`503 not-ready`: unverified bundle, unready worker, unloaded plugin. No `Retry-After`, deliberately."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: The one error envelope, every plane (contracts §3.1). The code list is closed.
      additionalProperties: false
      required: [error, detail]
      properties:
        error:
          type: string
          enum:
            - bad-credential
            - expired-token
            - unknown
            - conflict
            - contract
            - backpressure
            - fail-closed
            - not-ready
        detail:
          type: string
          description: Human-readable. Identical for every branch of a `404`, so it discloses nothing an enumeration could use.
        retry_after_s:
          type: integer
          minimum: 0
          description: Present on every `429` and on nothing else; agrees with the `Retry-After` header.
    BackpressureError:
      description: An `Error` whose code is `backpressure` and whose `retry_after_s` is present.
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [retry_after_s]
          properties:
            error:
              const: backpressure
...
    ArtifactRequest:
      type: object
      additionalProperties: false
      required: [view]
      properties:
        view:
          type: string
          description: Which view's row space the count is taken in. Required, unlike `/v1/items` — a masked count is an intersection in row space, and row space is per view.
        idset:
          type: integer
          minimum: 0
        zoom:
          type: integer
          minimum: 0
          maximum: 16
          description: The depth the caller draws at, for the vertex rule a predicate or an authored shape is served under (`polygon-membership.md` §7.2). Absent serves the whole presimplified shape under the 2,048-vertex guard. A derived hull is unaffected.
    ArtifactResponse:
      type: object
      additionalProperties: false
      required: [layer, masked_count, rung]
      properties:
        layer:
          type: string
        key:
          type: string
          description: The publisher's own key, absent where none was supplied.
        masked_count:
          type: integer
          minimum: 0
          description: How many of this artifact's members **the asking principal** can see. Never the artifact's size.
        centroid:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: number }
          description: In the viewport's 32-bit-per-axis grid units, over the members this principal can see. Absent where the layer declares none.
        box:
          type: array
          minItems: 4
          maxItems: 4
          items: { type: integer, minimum: 0 }
          description: "`[min_x, min_y, max_x, max_y]`, grid units."
        shape:
          type: array
          description: "**The artifact's one drawn geometry** — parts, then rings, then vertices — of the kind `/v1/meta`'s `shape` names for its layer (`polygon-membership.md` §7.1): the derived hull, every α-group its own part; the predicate shape; or the authored one. A part's first ring is its outer and the rest are holes. Absent where the layer draws none."
          items:
            type: array
            description: One part — its outer ring, then its holes.
            items:
              type: array
              description: One closed ring in grid units; a derived hull's counter-clockwise from its lowest vertex.
              items:
                type: array
                minItems: 2
                maxItems: 2
                items: { type: integer, minimum: 0 }
        content:
          type: array
          items: { type: string }
          description: The publisher's supplied content, positional to the layer's declared `supplied_content` types. Absent where the layer declares none; never partial.
        rung:
          type: integer
          minimum: 0
          description: "The rung this artifact sits at — the declared level on a levelled layer, `0` on a treed or flat one (which, this response being one artifact with no parent links, is also its response-local chain depth; the viewport frame's `rung` column carries the same number). The one field here that is a fact about the artifact rather than about the asking principal: two principals served it agree on it."